Add constants file
