+++ docs/anchoring-assets/anchor-platform/getting_started.mdx
@@ -12,7 +12,7 @@ The easiest way to install the Anchor Platform is to pull the [docker image][anc
 <CodeExample>
 
 ```bash
-docker pull stellar/anchor-platform:2.0.1
+docker pull stellar/anchor-platform:[platform_version]
 ```
 
 </CodeExample>
@@ -30,7 +30,7 @@ Let's create a minimal compose file to get started.
 version: "3.8"
 services:
   sep-server:
-    image: stellar/anchor-platform:2.0.1
+    image: stellar/anchor-platform:[platform_version]
     command: --sep-server
     ports:
       - "8080:8080"
@@ -39,7 +39,7 @@ services:
     volumes:
       - ./config:/home
   platform-server:
-    image: stellar/anchor-platform:2.0.1
+    image: stellar/anchor-platform:[platform_version]
     command: --platform-server
     ports:
       - "8085:8085"
@@ -210,7 +210,7 @@ A database is only needed if using the Anchor Platform to facilitate transaction
 version: "3.8"
 services:
   sep-server:
-    image: stellar/anchor-platform:2.0.1
+    image: stellar/anchor-platform:[platform_version]
     command: --sep-server
     env_file:
       - ./dev.env
@@ -221,7 +221,7 @@ services:
     depends_on:
       - db
   platform-server:
-    image: stellar/anchor-platform:2.0.1
+    image: stellar/anchor-platform:[platform_version]
     command: --platform-server
     env_file:
       - ./dev.env
