+++ docs/anchoring-assets/anchor-platform/sep24/example.mdx
@@ -334,7 +334,7 @@ First, let's configure the Anchor Platform to observe the Stellar network for in
 # docker-compose.yml
 ---
 stellar-observer:
-  image: stellar/anchor-platform:2.0.1
+  image: stellar/anchor-platform:[platform_version]
   command: --stellar-observer
   env_file:
     - ./dev.env
