+++ docs/anchoring-assets/anchor-platform/sep31/integration.mdx
@@ -29,7 +29,7 @@ version: "3.8"
 
 services:
   sep-server:
-    image: stellar/anchor-platform:2.0.1
+    image: stellar/anchor-platform:[platform_version]
     command: --sep-server
     env_file:
       - ./dev.env
@@ -40,7 +40,7 @@ services:
     depends_on:
       - db
   platform-server:
-    image: stellar/anchor-platform:2.0.1
+    image: stellar/anchor-platform:[platform_version]
     command: --platform-server
     env_file:
       - ./dev.env
@@ -193,7 +193,7 @@ The Stellar Observer monitors the Stellar ledger for payments made to your accou
 services:
   ...
   observer:
-    image: stellar/anchor-platform:2.0.1
+    image: stellar/anchor-platform:[platform_version]
     command: --stellar-observer
     env_file:
       - ./dev.env
