+++ src/constants.ts
@@ -1,0 +1,27 @@
+export const PLATFORM_VERSION="2.1.1"
+
+export const CODE_LANGS = {
+    bash: 'bash',
+    cpp: 'C++',
+    curl: 'cURL',
+    dart: 'Flutter',
+    flutter: 'Flutter',
+    docker: 'Dockerfile',
+    go: 'Go',
+    html: 'html',
+    kotlin: 'Kotlin',
+    kt: 'Kotlin',
+    java: 'Java',
+    javascript: 'JavaScript',
+    js: 'JavaScript',
+    json: 'JSON',
+    json5: 'JSON5',
+    python: 'Python',
+    scss: 'SCSS',
+    sql: 'SQL',
+    toml: 'TOML',
+    ts: 'TypeScript',
+    tsx: 'TSX',
+    typescript: 'TypeScript',
+    yaml: 'YAML',
+};
